Week 06 Homework
Applied $.ready handler so all DOM elements would load first.

Created an array listing out all all the cities available for users to
select.

Added a ‘for’ loop  to add values to the dropdown menu using the items
from the city array.

used a .change handler to capture user actions so it recognised when a
user choses a new option.

If statements used to define how each option should interact with the
user.

applied .attr handler to add classes to change the background image
based on what city the user selects.

added a .removeAttr handler to remove previous class styles that may
have been added to change the background, and return the background
image to the default image.